Fixed first round of compilation issues
Locally, it seems like it should compile all the way but it doesn't link right because windows can't figure out whether it's 32-bit or 64-bit. Maybe it'll work on Github actions.

+++ load_opus.cpp
@@ -28,7 +28,7 @@ void load_opus(std::string const &filename, std::vector< float > *data_) {
 	//get length in samples:
 	ogg_int64_t length = op_pcm_total(op.get(), -1);
 	if (length >= 0) {
-		data.reserve(length);
+		data.reserve((unsigned int) length);
 	} else {
 		std::cerr << "WARNING: cannot estimate length of '" << filename << "', loading may be slow." << std::endl;
 		length = 0;
